fix warnings 5

+++ hyperdbg/hprdbghv/code/interface/Export.c
@@ -403,10 +403,10 @@ VmFuncGetInterruptibilityState()
 /**
  * @brief Clear STI and MOV SS bits
  *
- * @return UINT32
+ * @return UINT64
  */
-UINT32
-VmFuncClearSteppingBits(UINT32 Interruptibility)
+UINT64
+VmFuncClearSteppingBits(UINT64 Interruptibility)
 {
     return HvClearSteppingBits(Interruptibility);
 }
